Enum mmap::MapError

pub enum MapError {
    ErrFdNotAvail,
    ErrInvalidFd,
    ErrUnaligned,
    ErrNoMapSupport,
    ErrNoMem,
    ErrZeroLength,
    ErrUnknown(isize),
    ErrUnsupProt,
    ErrUnsupOffset,
    ErrAlreadyExists,
    ErrVirtualAlloc(i32),
    ErrCreateFileMappingW(i32),
    ErrMapViewOfFile(i32),
}

Possible errors when creating a map.

Variants
ErrFdNotAvail

The following are POSIX-specific

fd was not open for reading or, if using MapWritable, was not open for writing.

ErrInvalidFd

fd was not valid

ErrUnaligned

Either the address given by MapAddr or offset given by MapOffset was not a multiple of MemoryMap::granularity (unaligned to page size).

ErrNoMapSupport

With MapFd, the fd does not support mapping.

ErrNoMem

If using MapAddr, the address + min_len was outside of the process's address space. If using MapFd, the target of the fd didn't have enough resources to fulfill the request.

ErrZeroLength

A zero-length map was requested. This is invalid according to POSIX. Not all platforms obey this, but this wrapper does.

ErrUnknown(isize)

Unrecognized error. The inner value is the unrecognized errno.

ErrUnsupProt

The following are Windows-specific

Unsupported combination of protection flags (MapReadable/MapWritable/MapExecutable).

ErrUnsupOffset

When using MapFd, MapOffset was given (Windows does not support this at all)

ErrAlreadyExists

When using MapFd, there was already a mapping to the file.

ErrVirtualAlloc(i32)

Unrecognized error from VirtualAlloc. The inner value is the return value of GetLastError.

ErrCreateFileMappingW(i32)

Unrecognized error from CreateFileMapping. The inner value is the return value of GetLastError.

ErrMapViewOfFile(i32)

Unrecognized error from MapViewOfFile. The inner value is the return value of GetLastError.